requirements.txt
discord.py[voice]==2.4.0
discord-ext-voice-recv==0.5.2a179
openai==1.76.0
faster-whisper==1.1.1
edge-tts==6.1.12
python-dotenv==1.0.1
duckduckgo-search==6.3.0
aiohttp==3.11.18
pydantic==2.11.3
mss==9.0.1
pillow==10.4.0
requests==2.32.3
